Xcode 10 update, bump version to 1.1.0

+++ BadgeTableViewCell.m
@@ -49,7 +49,7 @@
         [self createBadge];
     }
 
-    badge.text = [NSString stringWithFormat:@"%d", badgeNumber];
+    badge.text = [NSString stringWithFormat:@"%d", (int)badgeNumber];
     badge.hidden = NO;
     [self layoutBadge];
 }
